+++ rs/index/src/optimizers/mod.rs
@@ -6,6 +6,6 @@ use quantization::quantization::Quantizer;
 
 use crate::segment::pending_segment::PendingSegment;
 
-pub trait SegmentOptimizer {
-    fn optimize<Q: Quantizer>(&self, segment: &PendingSegment<Q>) -> Result<()>;
+pub trait SegmentOptimizer<Q: Quantizer + Clone> {
+    fn optimize(&self, segment: &PendingSegment<Q>) -> Result<()>;
 }
